Driver acceptance ride

+++ App_Tim17/app/src/main/java/com/example/app_tim17/model/response/ride/Ride.java
@@ -3,10 +3,6 @@ package com.example.app_tim17.model.response.ride;
 
 import java.util.List;
 
-import com.example.app_tim17.model.response.Location;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
@@ -18,17 +14,13 @@ public class Ride {
     private Long id;
     @SerializedName("startTime")
     @Expose
-//    @JsonDeserialize(using = StringDeserializer.class)
-//    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
     private String startTime;
     @SerializedName("endTime")
     @Expose
-//    @JsonDeserialize(using = StringDeserializer.class)
-//    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
     private String endTime;
     @SerializedName("totalCost")
     @Expose
-    private Long totalCost;
+    private Integer totalCost;
     @SerializedName("driver")
     @Expose
     private DriverRideResponse driver;
@@ -37,7 +29,7 @@ public class Ride {
     private List<PassengerRideResponse> passengers = null;
     @SerializedName("estimatedTimeInMinutes")
     @Expose
-    private Long estimatedTimeInMinutes;
+    private Integer estimatedTimeInMinutes;
     @SerializedName("vehicleType")
     @Expose
     private String vehicleType;
@@ -57,7 +49,7 @@ public class Ride {
     public Ride() {
     }
 
-    public Ride(Long id, String startTime, String endTime, Long totalCost, DriverRideResponse driver, List<PassengerRideResponse> passengers, Long estimatedTimeInMinutes, String vehicleType, Boolean babyTransport, Boolean petTransport, RejectionResponse rejectionResponse, List<LocationForRide> locations) {
+    public Ride(Long id, String startTime, String endTime, Integer totalCost, DriverRideResponse driver, List<PassengerRideResponse> passengers, Integer estimatedTimeInMinutes, String vehicleType, Boolean babyTransport, Boolean petTransport, RejectionResponse rejectionResponse, List<LocationForRide> locations) {
         super();
         this.id = id;
         this.startTime = startTime;
@@ -97,11 +89,11 @@ public class Ride {
         this.endTime = endTime;
     }
 
-    public Long getTotalCost() {
+    public Integer getTotalCost() {
         return totalCost;
     }
 
-    public void setTotalCost(Long totalCost) {
+    public void setTotalCost(Integer totalCost) {
         this.totalCost = totalCost;
     }
 
@@ -121,11 +113,11 @@ public class Ride {
         this.passengers = passengers;
     }
 
-    public Long getEstimatedTimeInMinutes() {
+    public Integer getEstimatedTimeInMinutes() {
         return estimatedTimeInMinutes;
     }
 
-    public void setEstimatedTimeInMinutes(Long estimatedTimeInMinutes) {
+    public void setEstimatedTimeInMinutes(Integer estimatedTimeInMinutes) {
         this.estimatedTimeInMinutes = estimatedTimeInMinutes;
     }
